data_utils.py added for dates handling
